Uploaded excel_dashboard_project.xlsx file
Uploaded excel_dashboard_project.xlsx file containing the bike_buyers, working_sheet, pivot_table and dashboard sheets. The file contains the cleaned data set on the working_sheet sheet and on the pivot_table sheet it contains both pivot tables and pivot charts used for the analysis, the pivot charts are then copied onto the dashboard  sheet, which contains slicers (filtering toggles) that are linked to the pivot tables via Report Connections, these can be used to filter the data based upon criteria and can enable the reader to more quickly understand the correlations and patterns within the data. Processes involved in the cleaning included standardisation (I used filters to check for consistencies), I removed duplicates, corrected the currency formatting and changed the columns for age to age brackets to make the data more readable, I also changed the gender values from "M" and "F" to Male and Female to ensure that the charts and pivot tables had a more readable output. I changed the ages to age brackets using IF functions within IF functions and changed the gender column using Find and Replace. I also changed the commute distance value of "10+ Miles" to "More Than 10 Miles" because the data then was being sorted incorrectly and utilising letters enabled the "More Than 10 Miles" values to be the last column in a Data Table.

For more a more detailed breakdown of the project, please see the interactive README.md file within the repository.